+++ kit/src/main/java/io/github/dornol/excelkit/excel/TemplateListWriter.java
@@ -256,7 +256,8 @@ public class TemplateListWriter<T> {
         return new ExcelColumn<>(name, function, style, dataType.getSetter(),
                 c.minWidth, c.maxWidth, c.fixedWidth, c.dropdownOptions,
                 c.cellColorFunction, c.groupName, c.outlineLevel,
-                c.commentFunction, c.borderStyle, c.locked, c.hidden, c.validation);
+                c.commentFunction, c.borderStyle, c.locked, c.hidden, c.validation,
+                c.headerFontColor);
     }
 
     /**